Refactor Nano Benchmark (#4845)
* add benchmark check

* fix style

* save changes

* fix error

* add license and add ipex

* fix test

* fix style
