+++ .github/workflows/nano_unit_tests.yml
@@ -299,3 +299,34 @@ jobs:
           $CONDA/bin/conda remove -n openvino-tensorflow --all
         env:
           ANALYTICS_ZOO_ROOT: ${{ github.workspace }}
+
+  nano-unit-benchmark-sanity-check:
+    # The type of runner that the job will run on
+    runs-on: ${{ matrix.os }}
+    timeout-minutes: 20
+    strategy:
+      fail-fast: false
+      matrix:
+        os: ["ubuntu-20.04"]
+        python-version: ["3.7"]
+
+    # Steps represent a sequence of tasks that will be executed as part of the job
+    steps:
+      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
+      - uses: actions/checkout@v2
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install --upgrade setuptools==58.0.4
+          python -m pip install --upgrade wheel
+
+      - name: Run benchmark sanity check
+        shell: bash
+        run: bash python/nano/benchmark/start_benchmark.sh
+        env:
+          ANALYTICS_ZOO_ROOT: ${{ github.workspace }}
+          SANITY_CHECK: "1"
